Add a ListenerPool process to keep track of our listening sockets.
The ListenerPool process acts as a pool of sockets listening on the ports
specified in the application configuration under the :ports key. They can be
pulled from the pool to be used, and as long as they are out of the pool
cannot be pulled again. Once they are no longer being used, they should be
returned to the pool so that other processes can use them.

This will act as a source for listener sockets to use when starting test
servers.

Tests are included for the ListenerPool module. Since the ListenerPool process
is normally started as a supervised child of the application it is necessary
to stop the application in order to run the tests.

We also include a few helpers to get the list of ports from the application
configuration and to check if a given term refers to a socket.