Move the MockServer application module to MockServer.App.
We don't need to keep the Application code in the MockServer root module.
Since it really only is keeping track of our process tree, we move it down to
MockServer.App. We update the mix file to reflect the new application module,
and adjust our tests to reflect the change.