Add a parser for the mock data.
The mock server will need to get its data from somewhere, and that will
usually be a mock file. We add a parser for mock file data, as well as a
function to load the file from the mock path. The file format is also
described in the module documentation.

Tests are included for the parser, including basic negative tests.