Add support for comments in mock files.
We add the ability to include full line comments (beginning with a '#'
character) to a mock file. Empty lines are also ignored.

At the same time, we also fix a bug and get rid of an RE. The bug had to do
with parsing base64 encoded data which did not end with a '.' when the end of
file was encountered. This now generates a '{:error, :bad_base64, lineno}'
error.

The RE is removed because it was just being used to validate data. Instead we
let the String.to_integer/2 function fail, and rescue the failure.

We also make a change to the test code to provide slightly more idiomatic
examples of the lines being fed to the parser.

Additional assertions are added to verify the correct parsing of comments.