Let the MockData.load/1 function return a stream instead of a list.
Since the MockServer will be looking for a stream of {direction, data} tuples
for processing, we modify the load/1 function to return a stream instead of a
list. This will also permit a user defined stream of {direction, data} tuples
to be used by the server, as well as streams generated by load/1.

The tests are updated to reflect this change.