Clean up the MockData.parse/1 function, and tidy up the documentation.
The parse/1 function gets overhauled yet again. We go back to returning a
simple list of {:server | :client, data} tuples, as we're going to let users
who want to send unusually large datasets deliver their own tuples to the mock
data feeder, as the .mock file format is really designed for smaller datasets.

We also change the type of data parse/1 accepts to take a binary instead of a
list of lines. This is simpler from a number of perspectives.

Furthermore, the parse/1 function only returns the first error it encounters.
This makes sense for a number of reasons, and should make some of our other
code clearer.

The mock_path/0 and pathname/1 functions remain, and work much like before,
but the load/1 function gets removed as it will live in a different module.

The tests are updated to reflect the changes to the API. Also, we tweak the
test data to check that our error line numbers are sensitive to comments.