Add the MockDataFeed module to the application.
The MockDataFeed will act as a stateful queue delivering MockData messages to
the MockServer. It includes methods to start the process (start_link/0), load
the feed with mock data (load/2), and pull messages for consumption by the
MockServer (pull/1). It also contains a dump/1 method to help with any
debugging and/or testing.

Tests are included to verify operation.