Integrate MockDataFeed processes into the OTP process tree.
We'll be using MockDataFeed processes to feed mock data to MockServer
processes. As such, they should be integrated into the OTP process tree. We
put them under a supervisor which will not attempt to restart them if they
crash (which they generally shouldn't when working from a parsed file).

We also tweak some of the types and variable names in the MockDataFeed module,
but don't change the actual working for the feed processes.