Add the MockServer module to start a mock server using mock data.
The MockServer module is where the application has its root. This is where we
start a mock server process, which will process the data from a mock file (or
binary or separately managed mock data feed) and send it to a client over a
network connection, verifying that the client sends back the expected
responses.

The test code for this module is fairly minimal. We should really provide some
additional tests to ensure that we can run multiple servers simultaneously,
that the timeouts work as expected, and that errors cause the server to crash.
The simple positive test works fine, however.

One extra change made is the removal of the FeedSupervisor. We don't need it
any longer since MockDataFeeds are now linked to the MockServer process, which
is kept in the OTP tree by the ServerSupervisor.