Fix a bug in the tests.
A bug existed in the tests which would leave the mock_server application
stopped after the AppTest module. This would cause any later tests which
expected it to be running to fail (most notably the MockServerTest). Make sure
the application is running when this test module is entered and exited by
using the setup_all/2 macro and on_exit/1 function.