Enable MockServers to use multiple IP addresses.
Instead of constraining MockServers to run on 127.0.0.1 (IPv4 localhost), we
enable them to run on multiple, alternate IP interfaces. The primary
motivation was to enable running on IPv6, but this ability could also be used
to run on other local interfaces, or potentially even across systems.

This change requires a change to the MockServer API, as MockServer.start() now
requires an address to be specified in addition to a mock data source. As a
result, we bump the version number to 0.2.0.

We also add an additional module to simplify the logic of the ListenerPool.
The Listener module holds the state for a listener socket, including whether
or not it has been bound to a MockServer. This allows the ListenerPool state
to be immutable, as the required mutability is constrained to the Listener
processes in the pool. The result is a cleaner set of code.

Tests are updated to test the new module and exercise the changes in existing
APIs.

Finally, we increase the application requirements to Elixir 1.1.